refactor: apply well-known patterns — index types, async filter, interface cleanup
- GetEvent/GetItem now take size_t (was int): removes 28 redundant
  static_cast<int> casts across 13 files; the wrong type was inherited
  from the original wxWidgets port and propagated everywhere
- Extract MainWindow::RunAsyncFilter(): deduplicate the
  m_filteringInProgress + ToggleProgressVisibility + setFuture boilerplate
  shared by ApplyExtendedFilters and ApplyActorFilter (Template Method)
- Remove WxWidgetsUtils.hpp from MainController: leftover wx dependency
  eliminated now that GetEvent takes size_t
- IDataParser.hpp: remove inline comment blocks that narrated what the
  code already said; keep only non-obvious semantics (move-last-observer)
- IModel.hpp: remove constructor/destructor debug logging and Doxygen
  noise; add [[nodiscard]] to GetItem; drop Logger.hpp dependency
- FilterManager::applyFilters and EventsContainer::GetEvent marked
  [[nodiscard]]

+++ src/application/filters/Filter.cpp
@@ -285,7 +285,7 @@ std::vector<unsigned long> Filter::applyToIndices(const std::vector<unsigned lon
         // Fallback to legacy logic
         std::vector<unsigned long> result;
         for (unsigned long index : inputIndices) {
-            const auto& event = model.GetItem(static_cast<int>(index));
+            const auto& event = model.GetItem(index);
             if (matches(event)) {
                 result.push_back(index);
             }
@@ -298,7 +298,7 @@ std::vector<unsigned long> Filter::applyToIndices(const std::vector<unsigned lon
     for (const auto& condition : conditions) {
         std::vector<unsigned long> new_temp;
         for (unsigned long index : temp) {
-            const auto& event = model.GetItem(static_cast<int>(index));
+            const auto& event = model.GetItem(index);
             if (condition.matches(event)) {
                 new_temp.push_back(index);
             }
